requirements.txt
discord.py
sqlite3
python-dotenv

pip install -r requirements.txt